web/pyproject.toml
[project]
name = "pubmarine-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "numpy>=2.3.1",
    "pydantic>=2.11.7",
    "pyserial>=3.5",
    "pyserial-asyncio>=0.6",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
